Create cocoon package and refactored imports, logging
