+++ cocoon/__init__.py
@@ -1,0 +1,20 @@
+# import sys
+# from pathlib import Path
+
+# ROOT_PATH = Path(__file__).resolve().parent
+# sys.path.insert(0, ROOT_PATH)
+
+from .cluster_service import compute_cluster
+from .embedding_service import create_embeddings, initialize_faiss_index_from_embeddings
+from .output_service import generate_output
+from .search_service import find_entity_relation_matches_and_cluster
+
+__all__ = [
+    "core",
+    "utils",
+    "compute_cluster",
+    "create_embeddings",
+    "initialize_faiss_index_from_embeddings",
+    "generate_output",
+    "find_entity_relation_matches_and_cluster",
+]
