+++ cocoon/core/embeddings/__init__.py
@@ -1,0 +1,4 @@
+from .base import Embeddings
+from .bedrock import BedrockEmbeddings
+
+__all__ = ["Embeddings", "BedrockEmbeddings"]
