+++ cocoon/core/llm/__init__.py
@@ -1,0 +1,4 @@
+from .base import LLM
+from .bedrock import BedrockLLM
+
+__all__ = ["LLM", "BedrockLLM"]
