+++ cocoon/core/llm/bedrock.py
@@ -2,9 +2,11 @@ import json
 from typing import Any, Dict, List, Optional, Tuple, Union
 
 import boto3
-from core.llm.base import LLM
-from core.messages import AIMessage, BaseMessage, HumanMessage, SystemMessage, _message_type_lookups
-from utils.logging import logger
+
+from cocoon.utils.logging import logger
+
+from ..messages import AIMessage, BaseMessage, HumanMessage, SystemMessage, _message_type_lookups
+from .base import LLM
 
 
 def _format_anthropic_messages(messages: List[BaseMessage]):
