+++ cocoon/core/messages/__init__.py
@@ -1,0 +1,8 @@
+from .ai import AIMessage
+from .base import BaseMessage
+from .human import HumanMessage
+from .system import SystemMessage
+
+_message_type_lookups = {"human": "user", "ai": "assistant"}
+
+__all__ = ["_message_type_lookups", "BaseMessage", "SystemMessage", "AIMessage", "HumanMessage"]
