+++ cocoon/core/messages/base.py
