+++ cocoon/core/messages/system.py
@@ -1,4 +1,4 @@
-from core.messages.base import BaseMessage
+from .base import BaseMessage
 
 
 class SystemMessage(BaseMessage):
