+++ cocoon/core/search/__init__.py
@@ -1,0 +1,3 @@
+from .faiss_search import FaissEmbeddingSearchManager
+
+__all__ = ["FaissEmbeddingSearchManager"]
