+++ cocoon/output_service.py
