+++ pyproject.toml
@@ -1,3 +1,19 @@
+[build-system]
+requires = ["setuptools", "setuptools-scm"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "cocoon"
+version = "0.0.1"
+authors = [
+    {name = "Zachary Huang", email = "zh2408@columbia.edu"},
+]
+description = "Cocoon is an open-source project that aims to free analysts from tedious data transformations with LLM."
+dependencies = ['boto3', 'faiss-cpu', 'fsspec', 'numpy', 'openpyxl', 'pandas', 's3fs']
+
+[tool.setuptools]
+packages = ["cocoon"]
+
 [tool.ruff]
 # Exclude a variety of commonly ignored directories.
 exclude = [
