--- src/__init__.py
@@ -1,5 +1,0 @@
-import sys
-from pathlib import Path
-
-ROOT_PATH = Path(__file__).resolve().parent
-sys.path.insert(0, ROOT_PATH)
