--- src/core/messages/__init__.py
@@ -1,8 +1,0 @@
-from core.messages.ai import AIMessage
-from core.messages.base import BaseMessage
-from core.messages.human import HumanMessage
-from core.messages.system import SystemMessage
-
-_message_type_lookups = {"human": "user", "ai": "assistant"}
-
-__all__ = ["_message_type_lookups", "BaseMessage", "SystemMessage", "AIMessage", "HumanMessage"]
