--- tests/__init__.py
@@ -1,5 +1,0 @@
-import sys
-from pathlib import Path
-
-REPO_ROOT = Path(__file__).resolve().parent.parent.joinpath("src")
-sys.path.insert(0, str(REPO_ROOT))
